android/apk/AuthCodeManager.service/requirements.txt
alabaster==0.7.12
asgiref==3.2.10
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
Cheetah3==3.2.6
click==7.1.2
cobbler==3.1.1
coverage==5.3
crypto==1.4.1
distro==1.5.0
Django==3.1.2
dnspython==1.15.0
dnspython3==1.15.0
docutils==0.16
Flask==1.1.2
future==0.18.2
idna==2.10
imagesize==1.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
ldap3==2.8.1
MarkupSafe==1.1.1
mod-wsgi==4.7.1
Naked==0.1.31
netaddr==0.8.0
packaging==20.4
pyasn1==0.4.8
pycrypt==0.7.2
pycrypto==2.6.1
Pygments==2.7.2
pymongo==3.11.0
pyparsing==2.4.7
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
shellescape==3.8.1
simplejson==3.17.2
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
tornado==6.1
urllib3==1.25.11
Werkzeug==1.0.1
